# review: perf: Precompile patterns for identifier checks