--- a/src/main/java/spoon/support/reflect/reference/CtReferenceImpl.java
+++ b/src/main/java/spoon/support/reflect/reference/CtReferenceImpl.java
@@ -24,7 +24,6 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import static spoon.reflect.path.CtRole.NAME;
@@ -32,9 +31,6 @@ import static spoon.reflect.path.CtRole.NAME;
 public abstract class CtReferenceImpl extends CtElementImpl implements CtReference, Serializable {
 
 	private static final long serialVersionUID = 1L;
-	private static final Pattern IS_ARRAY_OR_INSTANCE = Pattern.compile("\\[\\]|@");
-	private static final Pattern IS_INNER_OR_GENERIC = Pattern.compile("\\.|<|>");
-	private static final Pattern IS_INNER_OR_GENERIC_OR_EMPTY = Pattern.compile("\\d.*|<.*>|^.{0}$");
 
 	// See isKeyword for more information on keywords
 	private static final Collection<String> baseKeywords = fillWithBaseKeywords();
@@ -59,7 +55,7 @@ public abstract class CtReferenceImpl extends CtElementImpl implements CtReferen
 	@Override
 	public <T extends CtReference> T setSimpleName(String simplename) {
 		Factory factory = getFactory();
-		checkIdentiferForJLSCorrectness(simplename);
+		checkIdentifierForJLSCorrectness(simplename);
 		if (factory == null) {
 			this.simplename = simplename;
 			return (T) this;
@@ -102,22 +98,88 @@ public abstract class CtReferenceImpl extends CtElementImpl implements CtReferen
 		}
 		return false;
 	}
-	private void checkIdentiferForJLSCorrectness(String simplename) {
-		/*
-		 * At the level of the Java Virtual Machine, every constructor written in the Java programming language (JLS §8.8)
-		 * appears as an instance initialization method that has the special name <init>.
-		 * This name is supplied by a compiler. Because the name is not a valid identifier,
-		 * it cannot be used directly in a program written in the Java programming language.
-		 */
-		//JDTTreeBuilderHelper.computeAnonymousName returns "$numbers$Name" so we have to skip them if they start with numbers
-		//allow empty identifier because they are sometimes used.
-		if (!IS_INNER_OR_GENERIC_OR_EMPTY.matcher(simplename).matches()) {
-			//split at "<" and ">" because "Iterator<Cache.Entry<K,Store.ValueHolder<V>>>" submits setSimplename ("Cache.Entry<K")
-			String[] splittedSimplename = IS_INNER_OR_GENERIC.split(simplename);
-			if (checkAllParts(splittedSimplename)) {
-				throw new SpoonException("Not allowed javaletter or keyword in identifier found. See JLS for correct identifier. Identifier: " + simplename);
+
+	/*
+	 * This method validates the simplename.
+	 * spoon needs to allow more names that are allowed by the JLS, as
+	 * - array references have a name, e.g. int[], where [] would not be allowed normally
+	 * - ? is used as name for intersection types
+	 * - <init>, <clinit>, <nulltype> are used to represent initializers and the null reference
+	 * - anonymous/local classes start with numbers in spoon
+	 * - simple names of packages are just their names, but they may contain '.'
+	 * - the name can contain generics, e.g. List<String>[]
+	 */
+	private void checkIdentifierForJLSCorrectness(String simplename) {
+		if (isSpecialType(simplename)) return;
+		if (!checkAll(simplename)) {
+			throw new SpoonException("Not allowed javaletter or keyword in identifier found. See JLS for correct identifier. Identifier: " + simplename);
+		}
+	}
+
+	/*
+	 * returns true if the name is valid.
+	 * this splits up the string into parts that need to be JLS compliant.
+	 */
+	private boolean checkAll(String name) {
+		int i = 0;
+		// leading digits come from anonymous/local classes. Skip them
+		while (i < name.length() && Character.isDigit(name.charAt(i))) {
+			i++;
+		}
+		int start = i; // used to mark the beginning of a part
+		char expectNext = 0; // 0 = do not expect anything
+		for (; i < name.length(); i++) {
+			if (expectNext != 0) {
+				if (name.charAt(i) != expectNext) {
+					return false;
+				} else if (name.charAt(i) == expectNext) {
+					expectNext = 0; // reset
+					continue; // skip it, no further checks required
+				}
+			}
+			switch (name.charAt(i)) {
+				case '.':
+				case '<':
+				case '>':
+					// we scanned a word of valid java identifiers (see default case) until one
+					// of the special delimiting chars that are allowed in spoon
+					// now we just need to make sure it is not a keyword
+					if (isKeyword(name.substring(start, i))) return false; // keyword -> not allowed
+					start = i + 1; // skip this special char
+					break;
+				case '[':
+					expectNext = ']'; // next char *must* close
+					break;
+				default: // if we come across an illegal java identifier char here, it's not valid at all
+					if (start == i) { // first char of a part
+						if (!Character.isJavaIdentifierStart(name.charAt(i))) {
+							return false;
+						}
+					} else {
+						if (!Character.isJavaIdentifierPart(name.charAt(i))) {
+							return false;
+						}
+					}
+					break;
 			}
 		}
+		// make sure the end state is correct too
+		if (expectNext != 0) {
+			return false; // expected something that didn't appear anymore
+		}
+		// e.g. a name that only contains valid java identifiers will end up here (start will never be updated)
+		// and we still need to make sure it is not a keyword.
+		// as updating start uses i + 1, it might be out of bounds, so avoid SIOOBEs here
+		if (start < name.length()) {
+			return !isKeyword(name.substring(start));
+		}
+		return true;
+	}
+
+	private static boolean isSpecialType(String identifier) {
+		return identifier.isEmpty()
+				|| "?".equals(identifier) // is wildcard, used for intersection types
+				|| (identifier.startsWith("<") && identifier.endsWith(">"));
 	}
 
 	/**
@@ -138,29 +200,6 @@ public abstract class CtReferenceImpl extends CtElementImpl implements CtReferen
 				|| (complianceLevel >= 9 && java9Keywords.contains(simplename)));
 	}
 
-	private boolean checkAllParts(String[] simplenameParts) {
-		for (String simpleName:simplenameParts) {
-			//because arrays use e.g. int[] and @Number is used for instances of an object e.g. foo@1
-			simpleName = IS_ARRAY_OR_INSTANCE.matcher(simpleName).replaceAll("");
-			if (isWildCard(simpleName)) {
-				// because in intersection types a typeReference sometimes has '?' as simplename
-				return false;
-			}
-			if (isKeyword(simpleName) || checkIdentifierChars(simpleName)) {
-				return true;
-			}
-		}
-		return false;
-	}
-	private boolean checkIdentifierChars(String simplename) {
-		if (simplename.length() == 0) {
-			return false;
-		}
-		return (!Character.isJavaIdentifierStart(simplename.charAt(0)))
-			|| simplename.chars().anyMatch(letter -> !Character.isJavaIdentifierPart(letter)
-		);
-	}
-
 	private static Collection<String> fillWithBaseKeywords() {
 		// removed types because needed as ref: "int","short", "char", "void", "byte","float", "true","false","boolean","double","long","class", "null"
 		// in the method isKeyword, more keywords are added to the checks based on the compliance level
@@ -169,13 +208,5 @@ public abstract class CtReferenceImpl extends CtElementImpl implements CtReferen
 				"transient", "catch", "extends", "try", "final", "interface", "static", "finally", "volatile",
 				"const",  "native", "super", "while")
 				.collect(Collectors.toCollection(HashSet::new));
-	}
-
-	/**
-	 * checks if the input is a wildcard '?'. The method is not null safe.
-	 * @return boolean true is input wildcard, false otherwise
-	 */
-	private boolean isWildCard(String name) {
-		return name.equals("?");
 	}
 }