--- a/src/test/java/spoon/test/parent/ContractOnSettersParametrizedTest.java
+++ b/src/test/java/spoon/test/parent/ContractOnSettersParametrizedTest.java
@@ -125,7 +125,7 @@ public class ContractOnSettersParametrizedTest<T extends CtVisitable> {
 				// an empty package is merged with the existing one
 				// we have to give it a name
 				if (argument instanceof CtPackage) {
-					((CtPackage) argument).setSimpleName(argument.getShortRepresentation());
+					((CtPackage) argument).setSimpleName(t.getClass().getPackage().getName());
 				}
 
 				return argument;