@I-Al-Istannen and I started looking into ways to improve performance of https://github.com/I-Al-Istannen/JavadocApi.

When looking into profilings, I noticed that `Pattern.compile()` had a very high invocation count (in fact, it was the highest of all traced methods). I tracked down the callees and ended up in in methods of `CtReferenceImpl` where regular expressions were used in combination with `String#matches()`, `String#replaceAll()` and `String#split()`. Precompiling these patterns seems to have a very high impact in our case. Some of the comparisons we've done with and without this patch:

From @I-Al-Istannen with the linked JavadocApi:
```
/usr/lib/jvm/java-16-adoptopenjdk/bin/java -Xmx8G -jar target/JavadocApi.jar   1612,75s user 3,70s system 690% cpu 3:54,06 total
/usr/lib/jvm/java-16-adoptopenjdk/bin/java -Xmx8G -jar target/JavadocApi.jar   1042,61s user 3,32s system 673% cpu 2:35,21 total
```
(scanning JavaFX, upper is without patch, lower is with patch)

```
/usr/lib/jvm/java-16-adoptopenjdk/bin/java -Xmx8G -jar target/JavadocApi.jar   1276,71s user 11,40s system 424% cpu 5:03,47 total
/usr/lib/jvm/java-16-adoptopenjdk/bin/java -Xmx8G -jar target/JavadocApi.jar   978,21s user 10,06s system 452% cpu 3:38,48 total
```
(scanning OpenJDK 16, upper is without patch, lower is with patch)

With YourKit, I profiled the `CtGenerationTest`, this is a screenshot of the differences focused on the `checkIdentifierForJLSCorrectness` method:
![yourkit result](https://user-images.githubusercontent.com/11150076/127744445-19d70fc3-b3a9-4730-a9a0-d158d1eb6613.png)