@SirYwell 2021-08-14
I spend a little bit more time on this and rewrote the logic without any regular expressions. I wrote microbenchmarks with JMH (<https://github.com/SirYwell/spoon-benchmark>, can be run with `./gradlew jmh`).

I changed the `ContractOnSettersparametrizedTest`, as it set the simple name of a `CtPackage` to `spoon.support.reflect.declaration.CtPackageImpl@1` which isn't a valid identifier (and it's the only case were a `@` was inserted). I'm not sure if the change there is appropriate but it makes the logic a little bit easier.

In general, the new implementation is more restrictive (e.g. `<<>>...@@@` was allowed as simple name before).

I also tried to add dense documentation to avoid further confusion.